refactor: import of labels
