+++ src/kili/adapters/kili_api_gateway/label/types.py
@@ -6,7 +6,7 @@ from typing import Dict, Optional
 
 @dataclass
 class UpdateLabelData:
-    """LabelData data."""
+    """Data to update a label."""
 
     is_sent_back_to_queue: Optional[bool]
     json_response: Optional[Dict]
